requirements.txt
# Core dependencies
python-dotenv==1.0.0
requests==2.31.0

# YouTube
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1

# AI
google-generativeai==0.3.2
anthropic==0.39.0

# Notion
notion-client==2.2.1

# Supabase
supabase==2.3.0

# Cloud Functions
functions-framework==3.5.0
